requirements.txt
paho-mqtt==1.1
wheel==0.24.0
wit==1.1
